+++ src/utilities.ts
@@ -190,6 +190,24 @@ function getWordAtPosition(document, position) {
     return text;
 }
 
+async function createNamespaceFromDocumentIfNotExists(doc) {
+
+    if (getConnectedState()) {
+        let document = getDocument(doc);
+        if (document) {
+            let ns = getNamespace(document);
+            let client = getSession(getFileType(document));
+            if (client) {
+                let nsList = await client.listNamespaces([]);
+                if (nsList['ns-list'] && nsList['ns-list'].includes(ns)) {
+                    return;
+                }
+                await client.eval("(ns " + ns + ")").value;
+            }
+        }
+    }
+}
+
 function getDocument(document): vscode.TextDocument {
     if (document && document.hasOwnProperty('fileName')) {
         return document;
@@ -237,11 +255,20 @@ function getSession(fileType = undefined): NReplSession {
     }
 }
 
-function getConnectedState() {
-    return state.cursor.get('connected');
+function getLaunchingState() { 
+    return state.deref().get('launching');
 }
 
-function setConnectedState(value) {
+function setLaunchingState(value: any) {
+    vscode.commands.executeCommand("setContext", "calva:launching", Boolean(value));
+    state.cursor.set('launching', value);
+}
+
+function getConnectedState() { 
+    return state.deref().get('connected');
+}
+
+function setConnectedState(value: Boolean) {
     if(value) {
         vscode.commands.executeCommand("setContext", "calva:connected", true);
         state.cursor.set('connected', true);
@@ -252,10 +279,10 @@ function setConnectedState(value) {
 }
 
 function getConnectingState() {
-    return state.cursor.get('connecting');
+    return state.deref().get('connecting');
 }
 
-function setConnectingState(value) {
+function setConnectingState(value: Boolean) {
     if(value) {
         vscode.commands.executeCommand("setContext", "calva:connecting", true);
         state.cursor.set('connecting', true);
@@ -396,11 +423,14 @@ export {
     getNamespace,
     getStartExpression,
     getWordAtPosition,
+    createNamespaceFromDocumentIfNotExists,
     getDocument,
     getDocumentNamespace,
     getFileType,
     getFileName,
     getSession,
+    getLaunchingState,
+    setLaunchingState,
     getConnectedState,
     setConnectedState,
     getConnectingState,